requirements.txt
# Werkzeug is famous for breaking Flask so pin down a known working version
Werkzeug==2.1.2

# Build dependencies
Flask==2.1.2
Flask-RESTX==0.5.1
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
python-dotenv==0.20.0

# Runtime dependencies
gunicorn==20.1.0
honcho==1.1.0

# Code quality
pylint==2.14.0
flake8==5.0.4
black==22.3.0

# Testing dependencies
nose==1.3.7
pinocchio==0.4.3
factory-boy==2.12.0

# Behavior Driven Development
behave==1.2.6
selenium==4.1.0
# selenium==4.3.0
compare==0.2b0
requests==2.28.1

# Code coverage
coverage==6.3.2
codecov==2.1.12

# Utilities
httpie==3.2.1
